Go back to using the default plugin_dir location
The default plugin_dir is widely used in the ecosystem and it is
causing pain to diverge. There are issues with weave needing to use a
distro provided cni but this is likely a problem in weave. The cni
bits will continue to exist in /usr/libexec/cni for the time being
however.